CLI/requirements.txt
click
pyperclip
imageio[ffmpeg]
pillow